requirements.txt
django
djangorestframework
drf-spectacular
whitenoise
pyjwt
django-cors-headers
psycopg2-binary
django-environ
requests
gunicorn